Export simpleRequest.
Initially, I decided to not export simpleRequest because it is less safe
then pagedRequest. However, I have realized the fairly obvious fact that
simpleRequest is needed for all non-GET requests (so anything intending
to write anything needs access to simpleRequest).

Perhaps a rename of simpleRequest -> request and pagedRequest ->
getRequest would be abstracted, but for now the current naming works as
this is an internal module. Hence the difference between the two
functions needs to be known so that it is obvious that simpleRequest
shouldn't be used for GET requests.