For non-conduit API functions, call resourceT for user.
For functions like getGists, which are a trivial wrapper around a
conduit based function, or functions which have no conduit based
function because they lack pagination, there is little use for the user
to be forced to have to call runResourceT. Hence, it is helpful to
automatically call this, so these functions now have a "runResourceT $"
at the start.

This is a major backwards compatible to the API, but luckily its not
even close to release yet.