Add support for creation of Gists.
In addition to the creation of the createGist function, which actually
allows the creation of a Gist easily and programmatically, this commit
also addresses a few bugs necessary for its completion. First, a User is
an optional field of a Gist (it would be nice if this was documented in
the GitHub API), because Gists can be anonymous. Hence, the User needs
to be a Maybe type.

A bug in the serialization to JSON of GistCreate caused the description
field to be completely ignored as the object field had a typo.

Another bug in the serialization to JSON of GistCreate set the filename
of all files to "filename" rather than the actual filename (because a
String was used instead of the variable with the same name)